fix: rename api user to profile  (#7355)
