+++ pages/map.js
@@ -8,7 +8,7 @@ import PageHead from "@components/PageHead";
 import Page from "@components/Page";
 import Badge from "@components/Badge";
 import { getTags } from "./api/discover/tags";
-import { getUsers } from "./api/users";
+import { getUsers } from "./api/profiles";
 import config from "@config/app.json";
 
 //this is required as leaflet is not compatible with SSR
